Make config public in client and serializable
This makes it possible to read out the configuration from the client
later for persistence and removes the need to carry the config around in
addition to the client.

Also make the config itself serde serializable, useful for smaller
tools that doesn't have its own data or to embed in larger configs.